helpers/auth.ts
